Create Card Chex Privacy Policy
This page serves as the privacy policy for Card Chex, developed by Epic Appz. 
https://play.google.com/store/apps/details?id=com.Card_Chex
This policy is subject to change with future updates.
By using Card Chex, any person or entity who downloads or interacts with the app acknowledges and agrees to the terms and conditions set out in this policy.